refactor: add constants file, babel packages, gitignore file, helper file and index files
