+++ extra/patterns/index.js
@@ -1,0 +1,4 @@
+import drawEquilateralTriangle from "./equilateral";
+import drawDiamond from "./diamond";
+
+export { drawDiamond, drawEquilateralTriangle};
